Replace uses of Manifest with ClassTag / TypeTag
The Manifest class is part of the pre-2.10 reflection API, which has
been replaced with the newer ClassTag and TypeTag classes.

As well as offering a richer mirror-based API, the new classes also
seem to resolve a bug with reflection-based class loading in Manifest,
which interacted badly with Play's DEV mode hot reloading.

Fixes #79